AP-9669: Use double instead of float for all config objects
In some cases where the step size is very small, float precision lead to rounding problems.
E.g. new Float(3.0 + 0.0000001) returns 3.0

AP-9669 (DL Keras: Node Generation Spinner Widget Bugs)

+++ org.knime.dl.keras/src/org/knime/dl/keras/core/config/constraint/DLKerasMaxNormConstraint.java
@@ -59,7 +59,7 @@ import org.knime.dl.python.util.DLPythonUtils;
 public final class DLKerasMaxNormConstraint extends DLKerasAbstractAxisConstraint {
 
     @Parameter(label = "Maximum norm", min = "0.0000001", stepSize = "0.0000001")
-    private float m_maxValue = 2.0f;
+    private double m_maxValue = 2.0;
 
     /**
      */
